Refactored toplevel grounding.
Removed all toplevelgrounders. A grounder now knows whether it CAN (optimization) as if it is toplevel and can refrain from adding tseitins.
It works correctly on a few tests, but no unittests have been writting for it. It is only implemented to be used by boolform and quantforms atm.
Startheory has been removed from groundtheories, closetheory is still necessary.