Threading and lua fixup
Fixed the issue in which the lua code in interactive mode incorrectly parsed "::". This was because the lua execution was decoupled from the parsing and the translation to "." was in there.

Added basic threading support to the "executeprocedure" which runs lua commands. Here should be added clean signalhandling code in future to kill the thread when appropriate.