Existpredform bdds are working, clone NOT allowed on bddgeneration!
Clone is not allowed when generating bdds because there is no mapping being kept between cloned variables, which is necessary for the bdds to work. Refs .
The unit tests for generating exist predform bdds (also through the manager) and generating its values are working.
Bugfixes in enumloopkupgenerator and fobddfactory.
Added a generator test on enumlookupgenerator.