Finally, restartable parser.
After trying for a long time to get a c++ parser or a reentrant c parser (both failed), it turned out that a restart command exists for flex which resets all its globals. The c++ parser mainly failed because of issues with ylwrap compatibility. The reentrant c parser failed because its usage is quite obscure and we use a lot of the interface.