Refactored symbolic structure and propagatorfactory.
Added clone methods to fodot parse tree which do not clone variables.
Moved and check the code which generates a bdd given a formula and bounds on its symbols.
Moved the propagator creation out of the inferences, there is now code to generate a bdd propagator or a naive one (empty bdds for all symbols).