Lots of fixes
fixes #86: When searching for n models, if only m, m<n exist, a warning is given, but execution proceeds correctly
fixes #85: Due to tuples being added to relationships in which they did not fit
	fixed maketrue/makefalse/makeunknown to check for the types of their arguments and to check the size of the tuple being added
Moved limits inclusion to a separate file
	to cleanly handle double min issue
	to prevent more eclipse warnings
building was fixed to allow building static in debug also and corrected some output
varmapping bug was fixed by cleaning the varmapping after creating each grounder => issue was because of memory reuse for variables that had already been added
Fixed the tests which were still passing options as arguments
Found out that any domainelement that is input as an identifier is converted into a string. Checked for this and added it to the manual.
TODO: there seem to be an issue with make check in debug mode where 2 queens tests fail because of unsat, could not trace the reason.