Fixed multiple issues
Added extra debugging flag to cmaketests (_GLIBCXX_DEBUG), which does checking on stl container operations (slightly slows down debugging).
Fixed a minisatid issue found with this flag.
Googletest has to be build with the same flags as the executable, so I added it to the to-be-compiled libraries.
Fixed an issue which crashed some tests (seemingly random). Was caused by an over-the-end access on a vector due to an error in derivebounds. Valgrind could not find the error.
Removed all static id-generators from voc and structure, replaced with one id generator in globaldata (easier understandable debugging output).
Fixed the usage of ConjOrDisj datastructures during grounding. It is not checked at runtime whether the type (CONJ or DISJ) has efectively been set before called. Fixed numerous errors this way.