BUGFIX in cloning generators
The current usage of generators did not allow switching to another generator which generates the same variables.
As this will only happen during lazy grounding, support was added to clone those grounders (earlier) and to SET those variables AGAIN to the value of the last time the generator was run (but maybe overridden later by another one).
REMAINING ISSUE: if a generator is switched between a lazy grounding one and a normal one (can this happen?) the normal one will NOT reinit its values so will still contain errors!