Refactoring workflow:
Updated solver
Bugfix in runidp when whitespaces are present in the idp path.
Refactored and simplified grounderfactory. Less transformations are done and they are done later.
Some fixes to get it to compile with Clang3.2, but clang is still buggy.
Refactored the tests to have less duplication and allowing command specification in test().