Further refactored grounding
Grounding now changes its input, so the clones should happen externally and work further with those changes.
This has not yet been checked for anything but modelexpansion.
First examples work correctly using skolemization, except that it is not number-of-model-equivalent (which might be inherent, should be checked).
The structures now monitor the vocabulary. Changes in the vocabulary are reflected directly to the structure. Code can be simplified in places where this was necessary in the past.