Refactored grounding, bugfixed funcconstraints
Bugfix in funcconstraints: it is incorrect to add func constraints for all functions in a vocabulary. Indeed, in that case functions become target for e.g. calculate definitions, which is certainly not intended. TODO for calc definitions, it would in any case be more logical to project the final model onto the head vocabulary.
There is only one ground call any more, also containing the optimization whenever relevant. Cloning is done externally, grounding WILL change the theory.
ReplaceNestedWithTseitin has been simplified, but is not yet used.
Skolemization seems to behave correctly, but is currently disabled.