Fixes parsing issues, adds aggregate reification
Fixes #198: Insert segfaulted because a set was deleted bu the pointer was reused later on when an error was detected.

Adds an unnesting step for nested aggregates with cp which replaces them with a CONSTANT instead of a variable, such that only one formula needs to be grounded. TODO should be generalized when free variables are involved.

Fixes an issue where an empty domain was no checked by derivebounds and led to a crash.