Moved verbosity() method to local to grounders and mx
It used to be in globaldata, but the groundverbosity is now local to the grounders.
Also abstractgroundtheory now contains a proper method to get the right verbosity.
Doing this, every occurence of "Groundverbosity" has been replaced. Hence
I also removed the groundverbosity option.
This fixes #102. However, see #167, these changes have not been documented.