Refactored set rewriting in case of CP
Rewriting of a set for CP was done in the set grounder itself. This is incorrect as rewriting a set with conditions and terms to a term only set cannot be done independently of the aggregate, as the neutral element of the aggregate function is necessary.
The rewriting is now done by the aggregate grounders, saving both the conditions and the cpterms in a TsSet intermediately.