Improved propagation for the case where LUP=false
In order to get the previous commit do some useful stuff, the propagation should also run
when LUP is disabled. I added an extra parameter to choose to let the propagation run.

HOWEVER!!! This code is NOT YET CORRECT!  Since the propagationinfo is now passed to the generators
but NOT given to the structure, we can get a situation where some sentences are not grounded, but the
information that allows to do that is not in the structure.

We should think about how this can be fixed since running over the entire structure to do LUP might be too expensive.
I believe the solution will be in lazily running over the structure: every time we ground an atom, we first check if
it satisfies the ct or the cf bdd in the symbolic structure.