Fixed the bug with generation over builtins
Because of better propagation over builtins, introduced in 6d7a4a28ce0a,
there were more equalities in the BDDs. However this had as a result that
some generators could not be made.

The reason for this was that generating all x[nat] for which N(x) (N={1..3})
generated all natural numbers because N just returned a truegenerator (since it assumes its arguments are N, not nat)
This is now fixed