Improved generators over builtins
Did this by fixing part of a fixme in the code.
Solving is now also optimized for > and <, not only for =
There is still some part "fixme" left, which will be done later.

In the mean time, I created a test for this fix, a querytest.
This same test fixes #215 (since they were very related)