runidp.sh: don't run idp as a child
The purpose of runidp.sh is only to set up some environment variables
and then execute the actual idp binary.  There is no point in running
idp as a child as there is nothing to do after the child finishes.
On the other hand, running idp as a child has some significant drawbacks.
In particular, there is no portable way for a child to detect that its
parent has died.  If the parent dies for some reason (possibly at the
request of the user), then the child is kept running.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>