runidp: also call idp_rl_end when user presses ^D
idp_rl_end was only being called when the user explicitly typed "quit"
or "exit", but not when ^D is pressed.  Even though the current version
of idp_rl_end does not actually do anything, there appears to be no
reason why it should not be run when ^D is pressed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>