tests/mxtests/boundsNoLUP.cpp: use unique name for test
Commit b38421a (Added tests for mx with bounds but without LUP,
Fri Sep 14 12:12:48 2012 +0200) introduced a new test in boundsNoLUP.cpp
with the same name as a test in bounds.cpp.  This is not a problem on Linux
where each test file is compiled into a separate executable, but
it is a problem on other platforms, where they are combined into
a single executable.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>