Removed obsolete code and dangerous deletion
The deletion should not take place, since the "temp"
at that place is still in use as the domain of the connector
we are replacing the domain of the leaf here.

The deletion should only take place if we wanted to use
the equalities, which is not needed
(equalities are only useful in the direction
connector -> leaf, since only the connector can contain duplicate
variables.