Fixed issue with leafconnectors and duplicate variables
To clarify this case: suppose a leaf is P(x,x) and a connector
is P(a,b). lcd should contain all data linking these two predforms
we want that
leaftoconnector is x-> a or x-> b, doesn't matter which
and that equalities is a=b
and that connectortoleaf is
a->x AND b-> x
Here was the mistake. This was an OR in the previous implementation.

This fixes #272

Also added test for this behaviour that fails even when propagation
options are low