implemented merging sorts
Previously, when structures were merged, the sorts were left untouched.
This resulted in a behavior depending on the order we merged.
Now I implemented that sorts also get merged.
Different sorts can result in a behaviour where a relation was two-valued before merging, but
became threevalued after merging. (this is NOT new, see #270)
I resolved this by adding warnings whenever a relation is no longer two-valued because of
adding new elements to a sort.

This fixes #270 (which was in fact no bug)