Improved applying propagation to structure
I fixed the inter method in FOPropBDDDomainFactory to always keep its tables
as small as possible. (by choosing to keep the pf table if the ct table is too large)
This method was generalized and is available from MainStructureComponents.

I also added a test. However, this test will fail if bounds are disabled.
In order to use it, we need a better testframework, (see #299)
Thus: test is added in TODO