removed FalseQuantKernelGenerator
In the code, it was already mentioned that FalseQuantKernelGenerators are actually a
twochildgenerator. They were not smarter or more efficient than twochildgenerators
hence there was no reason why falsequantkernelgenerators should still exist.
Thus, I removed them. (the code as it was before was hard to debug).

Also: added more documentation with the changed code.