Added a check for free variables in term blocks.
In termblocks, free variables are not allowed.
This fixes #248, which had nothing to do with using
nested aggregates, but with the use of a domain element
inside a term. This domain element was treated as an
unquantified variable, and no error was thrown.

I also added a test to check for this behavior.