Memory fix in FuncInter
The problem was that, when a functinter was created with a functable,
the functable was stored on two locations. But, at later points
one of two could be deleted.
This is now fixed, by makeing two separate functables sharing the
same internalfunctable.
This fixes #311

Also adapted the mergetest to check this behavior