IMPORTANT: toString -> print!!!
Refactored the toString method for efficiency:
a print method returning a ToStream object
a templated method operator<<(stream, ToStream<T>)
with a default implementation using a put for objects and pointers (except when they are fundamental, then writing them directly).
which can be overridden for specialized behavior without a put method.

Advantages: no intermediate string construction + more logical naming

a toString method is still available which does generate the string first but so should be used only when the string itself is important!