Improved symbolic propagation a lot.
Previously, the propagatorfactory unnested a lot of domain terms that needed no unnesting:
As now is documented, we need to unnest domain terms from predicates
to make it possible to create leafconnectors afterwards.
However, since we do not create leafconnectors for built-in symbols, we
don't need to unnest domain terms from those symbols.
I disabled this unnesting. This change fixes issue #236

Added the two tests that were given in this specific issue.

this commit might be useful to check when solving #239