Fixed bug where ECNF printer segfaults when CP was on
Reason for the segfault is that the Structure::inter(..) procedure returned an
object that would be deleted later on. This has been circumvented by providing
a Structure::storableInter(..) procedure that does exactly the same, but it
returns a clone of the interpretation that is storable.

This is only necessary in the translateTerm() method in GroundTranslator.cpp,
because the sortable there might be needed later (e.g. for printing)

This fixes bug #353