Implemented own forall/2 rule to work around XSB's limitations
The standard Prolog forall/2 built-in is not designed to be used with tabling.
Errors "Illegal cut over incomplete tabled predicat X" are possible because
forall/2 is implemented as follows:

  forall(G1,G2) :- \+ ( call(G1), \+ call(G2) ).

And the "\+" operator uses cuts.

This can be circumvented by calling table_once(G2) instead of call(G2). From
the XSB manual:

  table_once(?Goal)
    "..."
    succeeds only once even if there are many solutions to the subgoal Pred.
    However, it does not "cut over" the subcomputation started by the subgoal
    Pred, thereby ensuring the correct evaluation of tabled subgoals.

Because the Prolog rules are constructed in such a way that the call G2 is
ground, we are also only interested in one solution to the call G2, making this
a very suitable alternative.