Only 'base' predicates are calculated
E.g. in the formula (! x : P(x) => Q(x)), only the symbols P_ct, P_cf, Q_ct and
Q_cf will be calculated (and not the tseitins of formulas). However, it is
still possible (and very likely) that in order to calculate these 'base'
predicates, the tseitin predicates need to be queried.

What we achieve with this is that no unground queries for tseitin predicates
are made when they can be avoided.