Improved handling of infinite ranges with XSB.
First: they are no longer printed as ranges. This implicitly assumes they are
also not called from any of the rules.
Second: Adaptations were made to the addition of instantiated variables. A list
of already instantiated variables is kept and it prevents from variables being
instantiated multiple times (needlessly). For the purpuse of infinite ranges,
it is momentarily assumed that they are generated by being the result of
operations. E.g. typing the following as part of a theory:

    x = 6/3.

Will result in the type "float" being derived for x. This is no problem because
the prolog code will be as follows:

    ... , IDPXSB_X is 6/3, ...

after which IDPXSB_X will be considered instantiated. Hence the variable with
an infinite derived type is instantiated.

This fixes bug #423