Fixed bug in Top-down rules for quantified formula's
The call to the subformula did not always have the right subterms. Example:

  ! x : ? y : P(x,y).

transformed into the following incorrect rule (with Tx the tseitin for the
forall formula):

  ! x y : P_ct(x,y) <- (!v : ( (v=y) | P_cf(v) ) ) & Tx_ct(x).

where the call to P_cf/2 lacked the variable x. The correct rule is:

  ! x y : P_ct(x,y) <- (!v : ( (v=y) | P_cf(x,v) ) ) & Tx_ct(x).

This fixes bug #432

Note that this fix has a serious impact on the running time of the propagation.
It may be desirable to ignore this rule altogether because of performance
reasons.