Added transformation to make definitions calculable
This transformation adds the rule "open_symbol(X) <- false." for every open
symbol that is not two-valued (given a structure).

This transformation is used to ensure that the propagation using the
approximating definition can be done using XSB. Every rule not present in the
definition has no possible information to add, so it is the right approach to
simply insert a rule with "symbol <- false" for those symbols. They will simply
not contribute towards any propagation.