Bugfix concerning aggregates
Added the compared term as an "input var to check" so that the variable is
ground before the comparison is made.

Also changed the order of the arguments in the comparison of the aggregate
result and the term.

It is important to put the term first, because the unnest aggs transformation
will rewrite any "AGG COMP TERM" into "TERM COMP* AGG" with COMP* the opposite
comparison operator of COMP
Thus making the term always the first argument of the comparison

This fixes both aspects of bug #457