Added support for comparison chains in XSB.
XSB now depends on the TheoryUtils to detect FormulaUtils as ranges and to
handle them correctly (get lower- and upperbound, get variable).

Added detection of one type of range:
	NUMBER (=)< VAR (=)< NUMBER
		or
	NUMBER >(=) VAR >(=) NUMBER
where NUMBER is an integer number and VAR has a sort that is a subtype of nat
or int.

This fixes some aspects of issue #402