Moved "undefined" check for Prolog queries
It used to be only checked for recursive calles in idpxsb_forall. Instead, the
check has been moved to the top-level query.

This means that all answers are generated and the "undefined" ones are left out
only at the end.

This bug is spotted by executing propagation using the approximating definition
for an NQueens instance of N=4 and the (1,2) square already filled with a
queen. Normally, propagation completely solves the problem, but with the
previous implementation this did not happen.