add initial tutorial
The tutorial is written from an outsider's point of view.
It currently lacks an introduction.
The current author is mentioned in the tutorial not to take any credit,
but to take the blame for any mistakes.

The examples have been tested with the current version of IDP,
but ideally they should be added to the test suite so that they
get tested automatically.