Simplified printing of structures
Removed the possibility to print <cf> and <u> depending on table sizes.
This was confusing for many users.

Maybe in the future we can make this kind of printing an option (if many people want it).
For now, if you have too large ct-tables, you can manually print the cf table.

Fixes #551