Removed detection of duplicate generation of facts
This would never happen but would filter out fact generation because the
_loaded list does not take arity into account.

E.g.: with a type T and predicate T(arg,arg...) both given.
      The predicate would not be generated

Even when using name() instead of nameNoArity() this leads to problems

E.g. when a predicate T(arg) and function T(arg):arg are given.
     One of the two would not be generated.

The filter of open symbols that would be already "loaded" is useless to
begin with: they are always only generated once. The _loaded list is only useful
for detecting which sorts are already generated as ranges (in the getRanges()
method).

This fixes issue #440.