See #616
This fixes a problem with twovaluedness of propagators.
The problem was that during creation, twovaluedness of function symbols was used, while at the end of the process, twovaluedness of PFSymbols.
They are not treated the same.
I extended the initial check to also take into account the graphed versions, and in the mean time, generalised the code.

In fact, this failing assertion was not even a real bug; code was correct without the assert.
The assert simply is an approx check that structure did not change. Now, it should always work