Fixed error when translating builtins
Sometimes (when generating approximating definitions, for example), builtins
would have the sorts of their arguments specified: e.g. =[int,int](var1,var2)

The current implementation of fqn_name of a symbol symply uses the toString.
This must be circumvented because any version of a built-in (with different
argument types) must be mapped to the same XSB built-in. E.g. both =[int,int]
and =[nat,nat] must map to "=" in XSB, hence when a built-in is detected,
the nameNoArity() is called instead of the fqn.

A separate XSB test for this was not added because currently any way to specify
the sorts of arguments of built-in is not supported (e.g. =[int,int](var1,var2)
cannot be parsed).