Added extra type check after each XSB predicate call
This is because any variable that was instantiated in that call might actually
range over a different/stricter domain.

This type check could be filtered out if one could detect what the types of the
predicate call were. Doing this at this stage would incur heavy refactoring,
since a PrologTerm has no notion of the type of it's called variables.

This fixes bug #642