Adjusted print to be able to work with negative integer domain elements.
Added a whitespace before and after each operator, since arguments may be of the
form -1. XSB only accepts this if they are written as (X = -1), and not if
written as (X=-1). In the latter case, XSB will throw an error when parsing the
file.

This fixes bug #678

Important: always add a whitespace before and after each operator, arguments may be of the form -1
which will only be accepted by XSB if they are written as (X = -1), and not if written as (X=-1)