Bugfix in deriveSorts:
Two problems are fixed here: first, during the loop, derivefuncs and derivepreds
should be ran AFTER the accept
As these two determine when to stop (i.e. _changed)

Now it was possible that the system stopped one iteration too early.

Second: _underivablevariables should be cleared every run. They are set during the top down phase and can never be removed from this set.

This fixes #683 for predicate symbols

For function symbols, there is a related issue, to be fixed soon