Refactored XSB translation of MINELEM and MAXELEM functions
They are now considered as "ordinary" interpreted symbols. Since their
interpretation is not that big (2 symbols per sort), they be added to the facts
sent to XSB, whether they are used or not.

This makes the code for transforming PredForms cleared, since a special check
need not be made whether it concerns a "MIN" or "MAX" predicate.

It is, however, necessary for these PFSymbols that they are translated using
their full FQN name. For this reason, the special check in the translator that
used to filter out all built-ins and used their nameNoArity() instead of the FQN
is adapted to only filter out the standard comparison predicates (=, <, >).