Removed incorrect assert that caused errors in debug mode
Internally, when graphing a function, the Function object is not
changed into a Predicate object. This is expected in the Assert
that was removed.

The resulting situation would be that if had an open that was a
function, the assert would be thrown, even if the function had
been graphed properly in the definition.

This fixes issue #806