fixed printing of sorts.
In common.hpp, two template functions are defined:
operator<<
and print

The first one calls the second.
The first was overridden for vectors, while the second was overriden locally (in insert.cpp)
Of course, this has no effect. It is the first that should have been overridden locally.

Now, sorts S in scope V are printed as V::S instead of (V,S) when generating error messages.